// src/utils/enums/balances.enum.ts
export enum BalanceAccountTypesEnum {
    default = 0,
    recopays = 1,
    payWing = 2,
    pay365 = 4,
    tappay = 5,
    firstPe = 7,
    xettle = 8,
    starpaisa = 10,
    finopay = 11,
    finopayStaticQR = 12,
    pay365New = 13,
    OFCPAY = 14,
    theUnionPay = 15,
    heksaPay = 17,
    letspe = 18,
    payTMe = 19,
}

export enum BalanceAccountTypesName {
    Recopays = 'Recopays',
    Xettle = 'Xettle',
    Starpaisa = 'Starpaisa',
    Finopay = 'Finopay',
    FinopayStaticQR = 'Finopay Static QR',
    Pay365 = 'Pay365',
    FirstPe = '1stPe',
    Tappay = 'Tappay',
    PayWing = 'PayWing',
    Default = 'Default',
    Pay365New = 'Pay365 New',
    OFCPAY = 'OFCPAY',
    theUnionPay = 'The UnionPay',
    HeksaPay = 'HEKSA Payment',
    LETSPE = 'LETSPE',
    PayTMe = 'PayTMe',
}

export enum BalanceTypeEnum {
    Recopays = 'Recopays',
    Payin = 'Payin',
    Payout = 'Payout',
    ManualPayout = 'ManualPayout',
    Settlement = 'Settlement',
    Opening = 'Opening',
    Xettle = 'Xettle',
    Starpaisa = 'Starpaisa',
    Finopay = 'Finopay',
    FinopayStaticQR = 'Finopay Static QR',
    OFCPAY = 'OFCPAY',
    theUnionPay = 'The UnionPay',
}